size

+++ main.py
@@ -304,7 +304,7 @@ def start_video_player():
             first_time_video_played = False
         #video_player.play_sync()
         if not max_people_reached():
-            duration_of_video = video_player.duration() * 1000
+            duration_of_video = int(video_player.duration() * 1000)
             root.after(duration_of_video,start_video_player)
     else:
         root.after(1000, check_usb_stick_exists)
